876 Poems Generated while Computer partially frozen
Launched generation porcess on Wednesday the 14th at 7:26pm.
On arriving in office at 2pm the next day: the G5 screen was dark, the
mouse visible and mobile, but the login screen did not appear. Rebooted
in safe mode,  ran first aid, machine seems to be ok but 6 weeks of
full time labour seem to have resulted in a mild fritz to the
read-write capabilities.